move wrap_string to StringOperations
