+++ tests/utils/test_wrap_string.py
@@ -3,7 +3,7 @@
 from tcex.utils import Utils
 
 
-class TestWrapText:
+class TestWrapString:
     """Test the TcEx Utils Module."""
 
     @staticmethod
@@ -23,3 +23,23 @@ class TestWrapText:
         answer = Utils.wrap_string(s, (' ', '|'))
 
         assert answer == '\n'.join(correct)
+
+    @staticmethod
+    def test_unwrappable():
+        """Test long string that can't be wrapped."""
+        # flake8: noqa: E501; pylint: disable=line-too-long
+        s = '# vv: Build|Difference|Duplicates|Fill|Intersection|Is In|Is Subset|Is Superset|Length|Pop|Reverse|Reverse Sort|Slice|Sort|Sort Lowercase|Symmetric Difference|Unique|Union'
+
+        answer = Utils.wrap_string(s, (','))
+
+        assert answer == s
+
+    @staticmethod
+    def test_empty_input():
+        """Test empty string input."""
+        assert Utils.wrap_string('') == ''
+
+    @staticmethod
+    def test_short_input():
+        """Test input shorter than length."""
+        assert Utils.wrap_string('foobar') == 'foobar'
